chore!: rename oAuth2 to OAuth2
 * refactor error handling
