+++ src/auth/index.ts
@@ -1,39 +1,33 @@
-import {eBayConfig} from '../types/apiTypes';
+import {eBayConfig} from '../types';
 import {createRequest, IEBayApiRequest} from '../request';
 import AuthNAuth from './authNAuth';
-import OAuth2 from './оAuth2';
+import OAuth2 from './oAuth2';
 
 /**
  * Container with Auth'N'Auth and OAuth2.
  */
 export default class Auth {
-    public readonly eBayConfig: eBayConfig;
-    public readonly req: IEBayApiRequest;
+  public readonly eBayConfig: eBayConfig;
+  public readonly req: IEBayApiRequest;
 
-    public readonly authNAuth: AuthNAuth;
-    public readonly oAuth2: OAuth2;
+  public readonly authNAuth: AuthNAuth;
+  // tslint:disable-next-line:variable-name
+  public readonly OAuth2: OAuth2;
 
-    constructor(config: eBayConfig, req = createRequest()) {
-        this.eBayConfig = config;
-        this.req = req;
+  constructor(config: eBayConfig, req = createRequest()) {
+    this.eBayConfig = config;
+    this.req = req;
 
-        this.authNAuth = new AuthNAuth(
-            this.eBayConfig,
-            this.req
-        );
+    this.authNAuth = new AuthNAuth(config, req);
+    this.OAuth2 = new OAuth2(config, req);
+  }
 
-        this.oAuth2 = new OAuth2(
-            this.eBayConfig,
-            this.req
-        );
+  public async getHeaderAuthorization(useIaf: boolean) {
+    if (this.authNAuth.eBayAuthToken) {
+      return 'Token ' + this.authNAuth.eBayAuthToken;
     }
 
-    public async getHeaderAuthorization(useIaf: boolean) {
-        if (this.authNAuth.eBayAuthToken) {
-            return 'Token ' + this.authNAuth.eBayAuthToken;
-        }
-
-        const accessToken = await this.oAuth2.getAccessToken();
-        return (useIaf ? 'IAF ' : 'Bearer ') + accessToken;
-    }
+    const accessToken = await this.OAuth2.getAccessToken();
+    return (useIaf ? 'IAF ' : 'Bearer ') + accessToken;
+  }
 }
