+++ src/types/apiTypes.ts
@@ -24,14 +24,10 @@ export type eBayConfig = Keyset & {
   acceptLanguage?: string
 };
 
-export type RequestConfig = {
+export type AppConfig = eBayConfig & {
+  autoRefreshToken?: boolean,
   axiosConfig?: AxiosRequestConfig
 };
-
-/**
- * Not eBay Config.
- */
-export type AppConfig = eBayConfig & RequestConfig;
 
 export type AuthToken = {
   eBayAuthToken: string,
